pyproject.toml
[project]
name = "bc-recommendation-agent"
version = "0.1.0"
description = "Birth Control Recommendation Agent — LangGraph-based agentic system"
requires-python = ">=3.11"

dependencies = [
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "langchain-anthropic>=0.2",
    "langchain-ollama>=0.1",
    "langchain-openai>=0.2",
    "httpx>=0.27",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "pydantic>=2.0",
    "pandas>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
